requirements.txt
boto3==1.24.10
botocore==1.27.10
click==8.1.3
Flask==2.1.2
Flask-Login==0.5.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.9
Werkzeug==2.0.0
